+++ sentry/src/main/java/io/sentry/DataCategory.java
@@ -1,9 +1,7 @@
 package io.sentry;
 
-import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
-@ApiStatus.Internal
 public enum DataCategory {
   All("__all__"),
   Default("default"), // same as Error