+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -6,6 +6,7 @@ import io.sentry.backpressure.NoOpBackpressureMonitor;
 import io.sentry.cache.IEnvelopeCache;
 import io.sentry.cache.PersistingScopeObserver;
 import io.sentry.clientreport.ClientReportRecorder;
+import io.sentry.clientreport.DiscardReason;
 import io.sentry.clientreport.IClientReportRecorder;
 import io.sentry.clientreport.NoOpClientReportRecorder;
 import io.sentry.internal.debugmeta.IDebugMetaLoader;
@@ -178,6 +179,9 @@ public class SentryOptions {
    * to the scope. When nothing is returned from the function, the breadcrumb is dropped
    */
   private @Nullable BeforeBreadcrumbCallback beforeBreadcrumb;
+
+  /** Invoked when some data from the SDK is dropped before being consumed by Sentry */
+  private @Nullable OnDiscardCallback onDiscard;
 
   /** The cache dir. path for caching offline events */
   private @Nullable String cacheDirPath;
@@ -902,6 +906,24 @@ public class SentryOptions {
    */
   public void setBeforeBreadcrumb(@Nullable BeforeBreadcrumbCallback beforeBreadcrumb) {
     this.beforeBreadcrumb = beforeBreadcrumb;
+  }
+
+  /**
+   * Returns the onDiscard callback
+   *
+   * @return the onDiscard callback or null if not set
+   */
+  public @Nullable OnDiscardCallback getOnDiscard() {
+    return onDiscard;
+  }
+
+  /**
+   * Sets the onDiscard callback
+   *
+   * @param onDiscard the onDiscard callback
+   */
+  public void setOnDiscard(@Nullable OnDiscardCallback onDiscard) {
+    this.onDiscard = onDiscard;
   }
 
   /**
@@ -2980,6 +3002,20 @@ public class SentryOptions {
      */
     @Nullable
     Breadcrumb execute(@NotNull Breadcrumb breadcrumb, @NotNull Hint hint);
+  }
+
+  /** The OnDiscard callback */
+  public interface OnDiscardCallback {
+
+    /**
+     * Best-effort record of data discarded before reaching Sentry
+     *
+     * @param reason the reason data was dropped
+     * @param category the type of data discarded
+     * @param number the number of discarded data items
+     */
+    void execute(
+        @NotNull DiscardReason reason, @NotNull DataCategory category, @NotNull Long number);
   }
 
   /** The traces sampler callback. */